requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0
lpips>=0.1.4
clip @ git+https://github.com/openai/CLIP.git
numpy>=1.24.4
opencv-python
Pillow
tqdm
datasets
